requirements.txt
argcomplete==2.0.0
asgiref==3.6.0
autopep8==2.0.1
awscli==1.27.53
Beaker==1.10.0
beautifulsoup4==4.11.1
blivet==3.5.0
blivet-gui==2.4.0
botocore==1.29.53
Brlapi==0.8.4
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
cryptography==37.0.2
cupshelpers==1.0
dasbus==1.6
dbus-python==1.3.2
distro==1.7.0
Django==4.1.7
django-livereload-server==0.4
djangorestframework==3.14.0
docker==6.0.1
docutils==0.18.1
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
git-filter-repo==2.34.0
gpg==1.17.0
humanize==3.13.1
idna==3.3
jmespath==1.0.0
langtable==0.0.61
libcomps==0.1.18
lxml==4.9.1
Mako==1.1.4
MarkupSafe==2.1.1
nftables==0.1
olefile==0.46
packaging==21.3
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==9.2.0
pip==22.2.2
ply==3.11
productmd==1.33
ptyprocess==0.6.0
pwquality==1.4.5
pyasn1==0.4.8
pycairo==1.21.0
pycodestyle==2.10.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.1
pyenchant==3.2.2
PyGObject==3.42.2
pykickstart==3.41
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyparted==3.12.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.50
pytz==2022.7.1
pyudev==0.23.2
pyxdg==0.27
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.18.0
rsa==4.8
s3transfer==0.6.0
selinux==3.4
sepolicy==3.4
setools==4.4.0
setuptools==62.6.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.4
soupsieve==2.3.2.post1
sqlparse==0.4.3
systemd-python==235
Tempita==0.5.2
tornado==6.2
urllib3==1.26.12
websocket-client==1.4.2
